+++ composeApp/src/commonMain/kotlin/com/korealm/aria/shared/AudioController.kt
@@ -7,6 +7,6 @@ import com.korealm.aria.model.AudioResource
 interface AudioController {
     suspend fun play(audio: AudioResource)
     suspend fun stop(audio: AudioResource)
-    suspend fun setVolume(audio: AudioResource, volume: Float)
-    suspend fun setGlobalVolume(volume: Float)
+    suspend fun setVolume(audio: AudioResource, volume: Double)
+    suspend fun setGlobalVolume(volume: Double)
 }
